An idea about the new project.
Here I am trying to keep the mail Session for a mailbox open and create
another only when we try to change the mailbox account. 
I created Credentials, a Credentials object holds all the necessary
fields required for login.
SessionManager is an object the can return a Session instance.  The
getSessionInstance(Credentials) gets a Credentials object as a parameter
and returns a new Session Instance only if the Credentials are different
from the current one. 
MailBoxHandle uses a session to do what is needed with the mailbox. 
 *  This is just an idea. Criticism is welcomed.